Move helper for listing CVR exports out of legacy_cast_vote_records.ts
And then delete legacy_cast_vote_records.ts
